+++ exo-go/src/cmd/create.go
@@ -2,11 +2,9 @@ package cmd
 
 import (
 	"fmt"
-	"os"
 
-	"github.com/Originate/exosphere/exo-go/src/template_helpers"
+	"github.com/Originate/exosphere/exo-go/src/exotemplate"
 	"github.com/spf13/cobra"
-	"github.com/tmrts/boilr/pkg/template"
 )
 
 var createCmd = &cobra.Command{
@@ -17,18 +15,11 @@ var createCmd = &cobra.Command{
 			return
 		}
 		fmt.Print("We are about to create a new Exosphere application\n\n")
-		templateDir, err := templateHelpers.CreateApplicationTemplateDir()
+		templateDir, err := exotemplate.CreateApplicationTemplateDir()
 		if err != nil {
 			panic(err)
 		}
-		applicationTemplate, err := template.Get(templateDir)
-		if err != nil {
-			panic(err)
-		}
-		if err = applicationTemplate.Execute("."); err != nil {
-			panic(err)
-		}
-		if err = os.RemoveAll(templateDir); err != nil {
+		if err := exotemplate.Run(templateDir, "."); err != nil {
 			panic(err)
 		}
 		fmt.Println("\ndone")
